AES encryption/requirements.txt
opencv-python==4.9.0.80
numpy==1.26.4
pycryptodome==3.20.0
scikit-image==0.22.0
